+++ test/Microsoft.AspNet.Razor.Test/TagHelpers/TagHelperParseTreeRewriterTest.cs
@@ -1209,322 +1209,6 @@ namespace Microsoft.AspNet.Razor.Test.TagHelpers
                 expectedErrors: Enumerable.Empty<RazorError>());
         }
 
-        public static TheoryData EmptyTagHelperBoundAttributeData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-                var emptyAttributeError =
-                    "Attribute '{0}' on tag helper element '{1}' requires a value. Tag helper bound attributes of " +
-                    "type '{2}' cannot be empty or contain only whitespace.";
-                var boolTypeName = typeof(bool).FullName;
-
-                // documentContent, expectedOutput, expectedErrors
-                return new TheoryData<string, MarkupBlock, RazorError[]>
-                {
-                    {
-                        "<myth bound='' />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", new MarkupBlock())
-                                })),
-                        new[]
-                        {
-                            new RazorError(
-                                string.Format(emptyAttributeError, "bound", "myth", boolTypeName),
-                                absoluteIndex: 6, lineIndex: 0, columnIndex: 6, length: 5)
-                        }
-                    },
-                    {
-                        "<myth bound='    true' />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", factory.CodeMarkup("    true"))
-                                })),
-                        new RazorError[0]
-                    },
-                    {
-                        "<myth bound='    ' />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", factory.CodeMarkup("    "))
-                                })),
-                        new[]
-                        {
-                            new RazorError(
-                                string.Format(emptyAttributeError, "bound", "myth", boolTypeName),
-                                absoluteIndex: 6, lineIndex: 0, columnIndex: 6, length: 5)
-                        }
-                    },
-                    {
-                        "<myth bound=''  bound=\"\" />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", new MarkupBlock()),
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", new MarkupBlock())
-                                })),
-                        new[]
-                        {
-                            new RazorError(
-                                string.Format(emptyAttributeError, "bound", "myth", boolTypeName),
-                                absoluteIndex: 6, lineIndex: 0, columnIndex: 6, length: 5),
-                            new RazorError(
-                                string.Format(emptyAttributeError, "bound", "myth", boolTypeName),
-                                absoluteIndex: 16, lineIndex: 0, columnIndex: 16, length: 5)
-                        }
-                    },
-                    {
-                        "<myth bound=' '  bound=\"  \" />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", factory.CodeMarkup(" ")),
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", factory.CodeMarkup("  "))
-                                })),
-                        new[]
-                        {
-                            new RazorError(
-                                string.Format(emptyAttributeError, "bound", "myth", boolTypeName),
-                                absoluteIndex: 6, lineIndex: 0, columnIndex: 6, length: 5),
-                            new RazorError(
-                                string.Format(emptyAttributeError, "bound", "myth", boolTypeName),
-                                absoluteIndex: 17, lineIndex: 0, columnIndex: 17, length: 5)
-                        }
-                    },
-                    {
-                        "<myth bound='true' bound=  />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", factory.CodeMarkup("true")),
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", factory.CodeMarkup(string.Empty).With(SpanChunkGenerator.Null))
-                                })),
-                        new[]
-                        {
-                            new RazorError(
-                                string.Format(emptyAttributeError, "bound", "myth", boolTypeName),
-                                absoluteIndex: 19, lineIndex: 0, columnIndex: 19, length: 5)
-                        }
-                    },
-                    {
-                        "<myth bound= name='' />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", factory.CodeMarkup(string.Empty).With(SpanChunkGenerator.Null)),
-                                    new KeyValuePair<string, SyntaxTreeNode>("name", new MarkupBlock())
-                                })),
-                        new[]
-                        {
-                            new RazorError(
-                                string.Format(emptyAttributeError, "bound", "myth", boolTypeName),
-                                absoluteIndex: 6, lineIndex: 0, columnIndex: 6, length: 5),
-                        }
-                    },
-                    {
-                        "<myth bound= name='  ' />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", factory.CodeMarkup(string.Empty).With(SpanChunkGenerator.Null)),
-                                    new KeyValuePair<string, SyntaxTreeNode>("name", factory.Markup("  "))
-                                })),
-                        new[]
-                        {
-                            new RazorError(
-                                string.Format(emptyAttributeError, "bound", "myth", boolTypeName),
-                                absoluteIndex: 6, lineIndex: 0, columnIndex: 6, length: 5),
-                        }
-                    },
-                    {
-                        "<myth bound='true' name='john' bound= name= />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", factory.CodeMarkup("true")),
-                                    new KeyValuePair<string, SyntaxTreeNode>("name", factory.Markup("john")),
-                                    new KeyValuePair<string, SyntaxTreeNode>("bound", factory.CodeMarkup(string.Empty).With(SpanChunkGenerator.Null)),
-                                    new KeyValuePair<string, SyntaxTreeNode>("name", factory.Markup(string.Empty).With(SpanChunkGenerator.Null))
-                                })),
-                        new[]
-                        {
-                            new RazorError(
-                                string.Format(emptyAttributeError, "bound", "myth", boolTypeName),
-                                absoluteIndex: 31, lineIndex: 0, columnIndex: 31, length: 5),
-                        }
-                    },
-                    {
-                        "<myth BouND='' />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("BouND", new MarkupBlock())
-                                })),
-                        new[]
-                        {
-                            new RazorError(
-                                string.Format(emptyAttributeError, "BouND", "myth", boolTypeName),
-                                absoluteIndex: 6, lineIndex: 0, columnIndex: 6, length: 5),
-                        }
-                    },
-                    {
-                        "<myth BOUND=''    bOUnd=\"\" />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("BOUND", new MarkupBlock()),
-                                    new KeyValuePair<string, SyntaxTreeNode>("bOUnd", new MarkupBlock())
-                                })),
-                        new[]
-                        {
-                            new RazorError(
-                                string.Format(emptyAttributeError, "BOUND", "myth", boolTypeName),
-                                absoluteIndex: 6, lineIndex: 0, columnIndex: 6, length: 5),
-                            new RazorError(
-                                string.Format(emptyAttributeError, "bOUnd", "myth", boolTypeName),
-                                absoluteIndex: 18, lineIndex: 0, columnIndex: 18, length: 5)
-                        }
-                    },
-                    {
-                        "<myth BOUND= nAMe='john'></myth>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("BOUND", factory.CodeMarkup(string.Empty).With(SpanChunkGenerator.Null)),
-                                    new KeyValuePair<string, SyntaxTreeNode>("nAMe", factory.Markup("john"))
-                                })),
-                        new[]
-                        {
-                            new RazorError(
-                                string.Format(emptyAttributeError, "BOUND", "myth", boolTypeName),
-                                absoluteIndex: 6, lineIndex: 0, columnIndex: 6, length: 5)
-                        }
-                    },
-                    {
-                        "<myth bound='    @true  ' />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    {
-                                        new KeyValuePair<string, SyntaxTreeNode>(
-                                            "bound",
-                                            new MarkupBlock(
-                                                new MarkupBlock(
-                                                factory.Markup("    "),
-                                                new ExpressionBlock(
-                                                    factory.CodeTransition(),
-                                                    factory.Code("true")
-                                                        .AsImplicitExpression(CSharpCodeParser.DefaultKeywords)
-                                                        .Accepts(AcceptedCharacters.NonWhiteSpace))),
-                                                factory.Markup("  ")))
-                                    }
-                                })),
-                        new RazorError[0]
-                    },
-                    {
-                        "<myth bound='    @(true)  ' />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock(
-                                "myth",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    {
-                                        new KeyValuePair<string, SyntaxTreeNode>(
-                                            "bound",
-                                            new MarkupBlock(
-                                                new MarkupBlock(
-                                                factory.Markup("    "),
-                                                new ExpressionBlock(
-                                                    factory.CodeTransition(),
-                                                    factory.MetaCode("(").Accepts(AcceptedCharacters.None),
-                                                    factory.Code("true").AsExpression(),
-                                                    factory.MetaCode(")").Accepts(AcceptedCharacters.None))),
-                                                factory.Markup("  ")))
-                                    }
-                                })),
-                        new RazorError[0]
-                    },
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(EmptyTagHelperBoundAttributeData))]
-        public void Rewrite_CreatesErrorForEmptyTagHelperBoundAttributes(
-            string documentContent,
-            MarkupBlock expectedOutput,
-            RazorError[] expectedErrors)
-        {
-            // Arrange
-            var descriptors = new TagHelperDescriptor[]
-                {
-                    new TagHelperDescriptor(
-                        tagName: "myth",
-                        typeName: "mythTagHelper",
-                        assemblyName: "SomeAssembly",
-                        attributes: new[]
-                        {
-                            new TagHelperAttributeDescriptor(
-                                name: "bound",
-                                propertyName: "Bound",
-                                typeName: typeof(bool).FullName,
-                                isIndexer: false),
-                            new TagHelperAttributeDescriptor(
-                                name: "name",
-                                propertyName: "Name",
-                                typeName: typeof(string).FullName,
-                                isIndexer: false)
-                        })
-                };
-            var descriptorProvider = new TagHelperDescriptorProvider(descriptors);
-
-            // Act & Assert
-            EvaluateData(descriptorProvider, documentContent, expectedOutput, expectedErrors);
-        }
-
         public static TheoryData OptOut_WithAttributeTextTagData
         {
             get
@@ -3204,1023 +2888,6 @@ namespace Microsoft.AspNet.Razor.Test.TagHelpers
             RunParseTreeRewriterTest(documentContent, expectedOutput, expectedErrors, "strong", "p");
         }
 
-        public static TheoryData EmptyAttributeTagHelperData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-
-                // documentContent, expectedOutput
-                return new TheoryData<string, MarkupBlock>
-                {
-                    {
-                        "<p class=\"\"></p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class",  new MarkupBlock())
-                            }))
-                    },
-                    {
-                        "<p class=''></p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class",  new MarkupBlock())
-                            }))
-                    },
-                    {
-                        "<p class=></p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                // We expected a markup node here because attribute values without quotes can only ever
-                                // be a single item, hence don't need to be enclosed by a block.
-                                new KeyValuePair<string, SyntaxTreeNode>(
-                                    "class",
-                                    factory.Markup("").With(SpanChunkGenerator.Null)),
-                            }))
-                    },
-                    {
-                        "<p class1='' class2= class3=\"\" />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("class1", new MarkupBlock()),
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "class2",
-                                        factory.Markup("").With(SpanChunkGenerator.Null)),
-                                    new KeyValuePair<string, SyntaxTreeNode>("class3", new MarkupBlock()),
-                                }))
-                    },
-                    {
-                        "<p class1=''class2=\"\"class3= />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("class1",  new MarkupBlock()),
-                                    new KeyValuePair<string, SyntaxTreeNode>("class2",  new MarkupBlock()),
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "class3",
-                                        factory.Markup("").With(SpanChunkGenerator.Null)),
-                                }))
-                    },
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(EmptyAttributeTagHelperData))]
-        public void Rewrite_UnderstandsEmptyAttributeTagHelpers(string documentContent, MarkupBlock expectedOutput)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, new RazorError[0], "p");
-        }
-
-        public static TheoryData<string, MarkupBlock, RazorError[]> MalformedTagHelperAttributeBlockData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-                var blockFactory = new BlockFactory(factory);
-                var errorFormatUnclosed = "Found a malformed '{0}' tag helper. Tag helpers must have a start and " +
-                                          "end tag or be self closing.";
-                var errorFormatNoCloseAngle = "Missing close angle for tag helper '{0}'.";
-                var errorFormatNoCSharp = "The tag helper '{0}' must not have C# in the element's attribute " +
-                                           "declaration area.";
-                Func<string, Block> createInvalidDoBlock = extraCode =>
-                {
-                    return new MarkupBlock(
-                        new MarkupBlock(
-                            new StatementBlock(
-                                factory.CodeTransition(),
-                                factory.Code("do {" + extraCode).AsStatement())));
-                };
-                var dateTimeNow = new MarkupBlock(
-                    new MarkupBlock(
-                        new ExpressionBlock(
-                            factory.CodeTransition(),
-                                factory.Code("DateTime.Now")
-                                    .AsImplicitExpression(CSharpCodeParser.DefaultKeywords)
-                                    .Accepts(AcceptedCharacters.NonWhiteSpace))));
-
-                return new TheoryData<string, MarkupBlock, RazorError[]>
-                {
-                    {
-                        "<p =\"false\"\" ></p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>())),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero)
-                        }
-                    },
-                    {
-                        "<p bar=\"false\"\" <strong>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("bar", factory.Markup("false"))
-                                })),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero)
-                        }
-                    },
-                    {
-                        "<p bar='false  <strong>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "bar",
-                                        new MarkupBlock(factory.Markup("false"), factory.Markup("  <strong>")))
-                                })),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero)
-                        }
-                    },
-                    {
-                        "<p bar='false  <strong'",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "bar",
-                                        new MarkupBlock(
-                                            factory.Markup("false"),
-                                            factory.Markup("  <strong")))
-                                })),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero)
-                        }
-                    },
-                    {
-                        "<p bar=false'",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "bar",
-                                        factory.Markup("false"))
-                                })),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                "TagHelper attributes must be well-formed.",
-                                absoluteIndex: 12,
-                                lineIndex: 0,
-                                columnIndex: 12)
-                        }
-                    },
-                    {
-                        "<p bar=\"false'",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "bar",
-                                        factory.Markup("false'"))
-                                })),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero)
-                        }
-                    },
-                    {
-                        "<p bar=\"false' ></p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "bar",
-                                        new MarkupBlock(
-                                            factory.Markup("false'"),
-                                            factory.Markup(" ></p>")))
-                                })),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero)
-                        }
-                    },
-                    {
-                        "<p foo bar<strong>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("foo", null),
-                                    new KeyValuePair<string, SyntaxTreeNode>("bar", null)
-                                },
-                                new MarkupTagHelperBlock("strong"))),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "strong"),
-                                absoluteIndex: 10, lineIndex: 0, columnIndex: 10)
-                        }
-                    },
-                    {
-                        "<p class=btn\" bar<strong>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("btn"))
-                                })),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero)
-                        }
-                    },
-                    {
-                        "<p class=btn\" bar=\"foo\"<strong>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("btn"))
-                                })),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero)
-                        }
-                    },
-                    {
-                        "<p class=\"btn bar=\"foo\"<strong>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "class",
-                                        new MarkupBlock(factory.Markup("btn"), factory.Markup(" bar="))),
-                                    new KeyValuePair<string, SyntaxTreeNode>("foo", null)
-                                },
-                                new MarkupTagHelperBlock("strong"))),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "strong"),
-                                absoluteIndex: 23, lineIndex: 0, columnIndex: 23)
-                        }
-                    },
-                    {
-                        "<p class=\"btn bar=\"foo\"></p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("class", new MarkupBlock(factory.Markup("btn"), factory.Markup(" bar="))),
-                                    new KeyValuePair<string, SyntaxTreeNode>("foo", null),
-                                })),
-                        new RazorError[0]
-                    },
-                    {
-                        "<p @DateTime.Now class=\"btn\"></p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p")),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCSharp, "p"),
-                                absoluteIndex: 3, lineIndex: 0 , columnIndex: 3)
-                        }
-                    },
-                    {
-                        "<p @DateTime.Now=\"btn\"></p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p")),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCSharp, "p"),
-                                absoluteIndex: 3, lineIndex: 0 , columnIndex: 3)
-                        }
-                    },
-                    {
-                        "<p class=@DateTime.Now\"></p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("class", dateTimeNow)
-                                })),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero)
-                        }
-                    },
-                    {
-                        "<p class=\"@do {",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("class", createInvalidDoBlock(string.Empty))
-                                })),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                RazorResources.FormatParseError_Expected_EndOfBlock_Before_EOF("do", "}", "{"),
-                                absoluteIndex: 11, lineIndex: 0, columnIndex: 11)
-                        }
-                    },
-                    {
-                        "<p class=\"@do {\"></p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("class", createInvalidDoBlock("\"></p>"))
-                                })),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                RazorResources.FormatParseError_Expected_EndOfBlock_Before_EOF("do", "}", "{"),
-                                absoluteIndex: 11, lineIndex: 0, columnIndex: 11),
-                            new RazorError(
-                                RazorResources.ParseError_Unterminated_String_Literal,
-                                absoluteIndex: 15, lineIndex: 0, columnIndex: 15)
-                        }
-                    },
-                    {
-                        "<p @do { someattribute=\"btn\"></p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p")),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCSharp, "p"),
-                                absoluteIndex: 3, lineIndex: 0 , columnIndex: 3),
-                            new RazorError(
-                                RazorResources.FormatParseError_Expected_EndOfBlock_Before_EOF("do", "}", "{"),
-                                absoluteIndex: 4, lineIndex: 0, columnIndex: 4),
-                            new RazorError(
-                                RazorResources.FormatParseError_UnexpectedEndTag("p"),
-                                absoluteIndex: 29, lineIndex: 0, columnIndex: 29)
-                        }
-                    }
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(MalformedTagHelperAttributeBlockData))]
-        public void Rewrite_CreatesErrorForMalformedTagHelpersWithAttributes(
-            string documentContent,
-            MarkupBlock expectedOutput,
-            RazorError[] expectedErrors)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, expectedErrors, "strong", "p");
-        }
-
-        public static TheoryData<string, MarkupBlock, RazorError[]> MalformedTagHelperBlockData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-                var blockFactory = new BlockFactory(factory);
-                var errorFormatUnclosed = "Found a malformed '{0}' tag helper. Tag helpers must have a start and " +
-                                          "end tag or be self closing.";
-                var errorFormatNoCloseAngle = "Missing close angle for tag helper '{0}'.";
-
-                return new TheoryData<string, MarkupBlock, RazorError[]>
-                {
-                    {
-                        "<p",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p")),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero)
-                        }
-                    },
-                    {
-                        "<p></p",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p")),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "p"),
-                                absoluteIndex: 3, lineIndex: 0, columnIndex: 3)
-                        }
-                    },
-                    {
-                        "<p><strong",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new MarkupTagHelperBlock("strong"))),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "strong"),
-                                absoluteIndex: 3, lineIndex: 0, columnIndex: 3),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "strong"),
-                                absoluteIndex: 3, lineIndex: 0, columnIndex: 3)
-                        }
-                    },
-                    {
-                        "<strong <p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("strong",
-                                new MarkupTagHelperBlock("p"))),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "strong"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "strong"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                absoluteIndex: 8, lineIndex: 0, columnIndex: 8)
-                        }
-                    },
-                    {
-                        "<strong </strong",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("strong")),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "strong"),
-                                SourceLocation.Zero),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatNoCloseAngle, "strong"),
-                                absoluteIndex: 8, lineIndex: 0, columnIndex: 8)
-                        }
-                    },
-                    {
-                        "<<</strong> <<p>",
-                        new MarkupBlock(
-                            blockFactory.MarkupTagBlock("<"),
-                            blockFactory.MarkupTagBlock("<"),
-                            blockFactory.MarkupTagBlock("</strong>"),
-                            factory.Markup(" "),
-                            blockFactory.MarkupTagBlock("<"),
-                            new MarkupTagHelperBlock("p")),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "strong"),
-                                absoluteIndex: 2, lineIndex: 0, columnIndex: 2),
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                absoluteIndex: 13, lineIndex: 0, columnIndex: 13)
-                        }
-                    },
-                    {
-                        "<<<strong>> <<>>",
-                        new MarkupBlock(
-                            blockFactory.MarkupTagBlock("<"),
-                            blockFactory.MarkupTagBlock("<"),
-                            new MarkupTagHelperBlock("strong",
-                                factory.Markup("> "),
-                                blockFactory.MarkupTagBlock("<"),
-                                blockFactory.MarkupTagBlock("<>"),
-                                factory.Markup(">"))),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "strong"),
-                                absoluteIndex: 2, lineIndex: 0, columnIndex: 2)
-                        }
-                    },
-                    {
-                        "<str<strong></p></strong>",
-                        new MarkupBlock(
-                            blockFactory.MarkupTagBlock("<str"),
-                            new MarkupTagHelperBlock("strong",
-                                blockFactory.MarkupTagBlock("</p>"))),
-                        new []
-                        {
-                            new RazorError(
-                                string.Format(CultureInfo.InvariantCulture, errorFormatUnclosed, "p"),
-                                absoluteIndex: 12, lineIndex: 0, columnIndex: 12)
-                        }
-                    }
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(MalformedTagHelperBlockData))]
-        public void Rewrite_CreatesErrorForMalformedTagHelper(
-            string documentContent,
-            MarkupBlock expectedOutput,
-            RazorError[] expectedErrors)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, expectedErrors, "strong", "p");
-        }
-
-        public static TheoryData CodeTagHelperAttributesData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-                var dateTimeNow = new MarkupBlock(
-                    factory.Markup(" "),
-                    new ExpressionBlock(
-                        factory.CodeTransition(),
-                            factory.Code("DateTime.Now")
-                                .AsImplicitExpression(CSharpCodeParser.DefaultKeywords)
-                                .Accepts(AcceptedCharacters.NonWhiteSpace)));
-
-                return new TheoryData<string, Block>
-                {
-                    {
-                        "<person age=\"12\" />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("person",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("age", factory.CodeMarkup("12"))
-                                }))
-                    },
-                    {
-                        "<person birthday=\"DateTime.Now\" />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("person",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "birthday",
-                                        factory.CodeMarkup("DateTime.Now"))
-                                }))
-                    },
-                    {
-                        "<person name=\"John\" />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("person",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("name", factory.Markup("John"))
-                                }))
-                    },
-                    {
-                        "<person name=\"Time: @DateTime.Now\" />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("person",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "name",
-                                        new MarkupBlock(factory.Markup("Time:"), dateTimeNow))
-                                }))
-                    },
-                    {
-                        "<person age=\"12\" birthday=\"DateTime.Now\" name=\"Time: @DateTime.Now\" />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("person",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("age", factory.CodeMarkup("12")),
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "birthday",
-                                        factory.CodeMarkup("DateTime.Now")),
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "name",
-                                        new MarkupBlock(factory.Markup("Time:"), dateTimeNow))
-                                }))
-                    },
-                    {
-                        "<person age=\"12\" birthday=\"DateTime.Now\" name=\"Time: @@ @DateTime.Now\" />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("person",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("age", factory.CodeMarkup("12")),
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "birthday",
-                                        factory.CodeMarkup("DateTime.Now")),
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "name",
-                                        new MarkupBlock(
-                                            factory.Markup("Time:"),
-                                             new MarkupBlock(
-                                                factory.Markup(" @").Accepts(AcceptedCharacters.None),
-                                                factory.Markup("@").With(SpanChunkGenerator.Null).Accepts(AcceptedCharacters.None)),
-                                            dateTimeNow))
-                                }))
-                    },
-                    {
-                        "<person age=\"12\" birthday=\"DateTime.Now\" name=\"@@BoundStringAttribute\" />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("person",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("age", factory.CodeMarkup("12")),
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "birthday",
-                                        factory.CodeMarkup("DateTime.Now")),
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "name",
-                                        new MarkupBlock(
-                                             new MarkupBlock(
-                                                factory.Markup("@").Accepts(AcceptedCharacters.None),
-                                                factory.Markup("@").With(SpanChunkGenerator.Null).Accepts(AcceptedCharacters.None)),
-                                            factory.Markup("BoundStringAttribute")))
-                                }))
-                    },
-                    {
-                        "<person age=\"@@@(11+1)\" birthday=\"DateTime.Now\" name=\"Time: @DateTime.Now\" />",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("person",
-                                selfClosing: true,
-                                attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "age",
-                                        new MarkupBlock(
-                                            new MarkupBlock(
-                                                factory.Markup("@").Accepts(AcceptedCharacters.None),
-                                                factory.Markup("@").With(SpanChunkGenerator.Null).Accepts(AcceptedCharacters.None)),
-                                            new MarkupBlock(
-                                                factory.EmptyHtml(),
-                                                new ExpressionBlock(
-                                                    factory.CodeTransition(),
-                                                    factory.MetaCode("(").Accepts(AcceptedCharacters.None),
-                                                    factory.Code("11+1").AsExpression(),
-                                                    factory.MetaCode(")").Accepts(AcceptedCharacters.None))))),
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "birthday",
-                                        factory.CodeMarkup("DateTime.Now")),
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "name",
-                                        new MarkupBlock(factory.Markup("Time:"), dateTimeNow))
-                                }))
-                    },
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(CodeTagHelperAttributesData))]
-        public void TagHelperParseTreeRewriter_CreatesMarkupCodeSpansForNonStringTagHelperAttributes(
-            string documentContent,
-            MarkupBlock expectedOutput)
-        {
-            // Arrange
-            var descriptors = new TagHelperDescriptor[]
-            {
-                new TagHelperDescriptor("person", "PersonTagHelper", "personAssembly",
-                    attributes: new[]
-                    {
-                        new TagHelperAttributeDescriptor("age", "Age", typeof(int).FullName, isIndexer: false),
-                        new TagHelperAttributeDescriptor(
-                            "birthday",
-                            "BirthDay",
-                            typeof(DateTime).FullName,
-                            isIndexer: false),
-                        new TagHelperAttributeDescriptor("name", "Name", typeof(string).FullName, isIndexer: false),
-                    })
-            };
-            var providerContext = new TagHelperDescriptorProvider(descriptors);
-
-            // Act & Assert
-            EvaluateData(providerContext,
-                         documentContent,
-                         expectedOutput,
-                         expectedErrors: Enumerable.Empty<RazorError>());
-        }
-
-        public static IEnumerable<object[]> IncompleteHelperBlockData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-                var blockFactory = new BlockFactory(factory);
-                var malformedErrorFormat = "Found a malformed '{0}' tag helper. Tag helpers must have a start and " +
-                                           "end tag or be self closing.";
-                var dateTimeNow = new MarkupBlock(
-                    new ExpressionBlock(
-                        factory.CodeTransition(),
-                            factory.Code("DateTime.Now")
-                                .AsImplicitExpression(CSharpCodeParser.DefaultKeywords)
-                                .Accepts(AcceptedCharacters.NonWhiteSpace)));
-
-                yield return new object[]
-                {
-                    "<p class=foo dynamic=@DateTime.Now style=color:red;><strong></p></strong>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                                new KeyValuePair<string, SyntaxTreeNode>("dynamic", new MarkupBlock(dateTimeNow)),
-                                new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                            },
-                            new MarkupTagHelperBlock("strong")),
-                            blockFactory.MarkupTagBlock("</strong>")),
-                    new RazorError[]
-                    {
-                        new RazorError(
-                            string.Format(CultureInfo.InvariantCulture, malformedErrorFormat, "strong"),
-                            absoluteIndex: 52, lineIndex: 0, columnIndex: 52),
-                        new RazorError(
-                            string.Format(CultureInfo.InvariantCulture, malformedErrorFormat, "strong"),
-                            absoluteIndex: 64, lineIndex: 0, columnIndex: 64)
-                    }
-                };
-                yield return new object[]
-                {
-                    "<div><p>Hello <strong>World</strong></div>",
-                    new MarkupBlock(
-                        blockFactory.MarkupTagBlock("<div>"),
-                        new MarkupTagHelperBlock("p",
-                            factory.Markup("Hello "),
-                            new MarkupTagHelperBlock("strong",
-                                factory.Markup("World")),
-                            blockFactory.MarkupTagBlock("</div>"))),
-                    new RazorError[]
-                    {
-                        new RazorError(
-                            string.Format(CultureInfo.InvariantCulture, malformedErrorFormat, "p"),
-                            absoluteIndex: 5, lineIndex: 0, columnIndex: 5)
-                    }
-                };
-                yield return new object[]
-                {
-                    "<div><p>Hello <strong>World</div>",
-                    new MarkupBlock(
-                        blockFactory.MarkupTagBlock("<div>"),
-                        new MarkupTagHelperBlock("p",
-                            factory.Markup("Hello "),
-                            new MarkupTagHelperBlock("strong",
-                                factory.Markup("World"),
-                                blockFactory.MarkupTagBlock("</div>")))),
-                    new RazorError[]
-                    {
-                        new RazorError(
-                            string.Format(CultureInfo.InvariantCulture, malformedErrorFormat, "p"),
-                            absoluteIndex: 5, lineIndex: 0, columnIndex: 5),
-                        new RazorError(
-                            string.Format(CultureInfo.InvariantCulture, malformedErrorFormat, "strong"),
-                            absoluteIndex: 14, lineIndex: 0, columnIndex: 14)
-                    }
-                };
-                yield return new object[]
-                {
-                    "<p class=\"foo\">Hello <p style=\"color:red;\">World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo"))
-                            },
-                            factory.Markup("Hello "),
-                            new MarkupTagHelperBlock("p",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                                },
-                                factory.Markup("World")))),
-                    new RazorError[]
-                    {
-                        new RazorError(
-                            string.Format(CultureInfo.InvariantCulture, malformedErrorFormat, "p"),
-                            SourceLocation.Zero)
-                    }
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(IncompleteHelperBlockData))]
-        public void TagHelperParseTreeRewriter_CreatesErrorForIncompleteTagHelper(
-            string documentContent,
-            MarkupBlock expectedOutput,
-            RazorError[] expectedErrors)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, expectedErrors, "strong", "p");
-        }
-
-        public static TheoryData<string, MarkupBlock> InvalidHtmlBlockData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-                var blockFactory = new BlockFactory(factory);
-                var dateTimeNow = new ExpressionBlock(
-                    factory.CodeTransition(),
-                        factory.Code("DateTime.Now")
-                            .AsImplicitExpression(CSharpCodeParser.DefaultKeywords)
-                            .Accepts(AcceptedCharacters.NonWhiteSpace));
-
-                return new TheoryData<string, MarkupBlock>
-                {
-                    {
-                        "<<<p>>></p>",
-                        new MarkupBlock(
-                            blockFactory.MarkupTagBlock("<"),
-                            blockFactory.MarkupTagBlock("<"),
-                            new MarkupTagHelperBlock("p",
-                                factory.Markup(">>")))
-                    },
-                    {
-                        "<<p />",
-                        new MarkupBlock(
-                            blockFactory.MarkupTagBlock("<"),
-                            new MarkupTagHelperBlock("p", selfClosing: true))
-                    },
-                    {
-                        "< p />",
-                        new MarkupBlock(
-                            new MarkupTagBlock(
-                                factory.Markup("<"),
-                                new MarkupBlock(
-                                    factory.Markup(" p")),
-                                factory.Markup(" />")))
-                    },
-                    {
-                        "<input <p />",
-                        new MarkupBlock(
-                            blockFactory.MarkupTagBlock("<input "),
-                            new MarkupTagHelperBlock("p", selfClosing: true))
-                    },
-                    {
-                        "< class=\"foo\" <p />",
-                        new MarkupBlock(
-                            new MarkupTagBlock(
-                                factory.Markup("<"),
-                                new MarkupBlock(
-                                    new AttributeBlockChunkGenerator(
-                                        name: "class",
-                                        prefix: new LocationTagged<string>(" class=\"", 1, 0, 1),
-                                        suffix: new LocationTagged<string>("\"", 12, 0, 12)),
-                                    factory.Markup(" class=\"").With(SpanChunkGenerator.Null),
-                                    factory.Markup("foo").With(new LiteralAttributeChunkGenerator(
-                                        prefix: new LocationTagged<string>(string.Empty, 9, 0, 9),
-                                        value: new LocationTagged<string>("foo", 9, 0, 9))),
-                                    factory.Markup("\"").With(SpanChunkGenerator.Null)),
-                                factory.Markup(" ")),
-                            new MarkupTagHelperBlock("p", selfClosing: true))
-                    },
-                    {
-                        "</<<p>/></p>>",
-                        new MarkupBlock(
-                            blockFactory.MarkupTagBlock("</"),
-                            blockFactory.MarkupTagBlock("<"),
-                            new MarkupTagHelperBlock("p",
-                                factory.Markup("/>")),
-                            factory.Markup(">"))
-                    },
-                    {
-                        "</<<p>/><strong></p>>",
-                        new MarkupBlock(
-                            blockFactory.MarkupTagBlock("</"),
-                            blockFactory.MarkupTagBlock("<"),
-                            new MarkupTagHelperBlock("p",
-                                factory.Markup("/>"),
-                                blockFactory.MarkupTagBlock("<strong>")),
-                            factory.Markup(">"))
-                    },
-                    {
-                        "</<<p>@DateTime.Now/><strong></p>>",
-                        new MarkupBlock(
-                            blockFactory.MarkupTagBlock("</"),
-                            blockFactory.MarkupTagBlock("<"),
-                            new MarkupTagHelperBlock("p",
-                                dateTimeNow,
-                                factory.Markup("/>"),
-                                blockFactory.MarkupTagBlock("<strong>")),
-                            factory.Markup(">"))
-                    },
-                    {
-                        "</  /<  ><p>@DateTime.Now / ><strong></p></        >",
-                        new MarkupBlock(
-                            blockFactory.MarkupTagBlock("</  "),
-                            factory.Markup("/"),
-                            blockFactory.MarkupTagBlock("<  >"),
-                            new MarkupTagHelperBlock("p",
-                                dateTimeNow,
-                                factory.Markup(" / >"),
-                                blockFactory.MarkupTagBlock("<strong>")),
-                            blockFactory.MarkupTagBlock("</        >"))
-                    },
-                    {
-                        "<p>< @DateTime.Now ></ @DateTime.Now ></p>",
-                        new MarkupBlock(
-                            new MarkupTagHelperBlock("p",
-                                new MarkupTagBlock(
-                                    factory.Markup("< "),
-                                    dateTimeNow,
-                                    factory.Markup(" >")),
-                                blockFactory.MarkupTagBlock("</ "),
-                                dateTimeNow,
-                                factory.Markup(" >")))
-                    }
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(InvalidHtmlBlockData))]
-        public void TagHelperParseTreeRewriter_AllowsInvalidHtml(string documentContent, MarkupBlock expectedOutput)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, "p");
-        }
-
         public static IEnumerable<object[]> TextTagsBlockData
         {
             get
@@ -4420,775 +3087,6 @@ namespace Microsoft.AspNet.Razor.Test.TagHelpers
             MarkupBlock expectedOutput)
         {
             RunParseTreeRewriterTest(documentContent, expectedOutput, "!--", "?xml", "![CDATA[", "!DOCTYPE");
-        }
-
-        public static IEnumerable<object[]> OddlySpacedBlockData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-
-                yield return new object[]
-                {
-                    "<p      class=\"     foo\"    style=\"   color :  red  ;   \"    ></p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                        new List<KeyValuePair<string, SyntaxTreeNode>>
-                        {
-                            new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("     foo")),
-                            new KeyValuePair<string, SyntaxTreeNode>(
-                                "style",
-                                new MarkupBlock(
-                                    factory.Markup("   color"),
-                                    factory.Markup(" :"),
-                                    factory.Markup("  red"),
-                                    factory.Markup("  ;"),
-                                    factory.Markup("   ")))
-                        }))
-                };
-                yield return new object[]
-                {
-                    "<p      class=\"     foo\"    style=\"   color :  red  ;   \"    >Hello World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("     foo")),
-                                new KeyValuePair<string, SyntaxTreeNode>(
-                                    "style",
-                                    new MarkupBlock(
-                                        factory.Markup("   color"),
-                                        factory.Markup(" :"),
-                                        factory.Markup("  red"),
-                                        factory.Markup("  ;"),
-                                        factory.Markup("   ")))
-                            },
-                            factory.Markup("Hello World")))
-                };
-                yield return new object[]
-                {
-                    "<p     class=\"   foo  \" >Hello</p> <p    style=\"  color:red; \" >World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>(
-                                    "class",
-                                    new MarkupBlock(factory.Markup("   foo"), factory.Markup("  ")))
-                            },
-                            factory.Markup("Hello")),
-                        factory.Markup(" "),
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>(
-                                    "style",
-                                    new MarkupBlock(factory.Markup("  color:red;"), factory.Markup(" ")))
-                            },
-                            factory.Markup("World")))
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(OddlySpacedBlockData))]
-        public void TagHelperParseTreeRewriter_RewritesOddlySpacedTagHelperTagBlocks(
-            string documentContent,
-            MarkupBlock expectedOutput)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, "p");
-        }
-
-        public static IEnumerable<object[]> ComplexAttributeTagHelperBlockData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-                var blockFactory = new BlockFactory(factory);
-                var dateTimeNowString = "@DateTime.Now";
-                var dateTimeNow = new MarkupBlock(
-                                    new ExpressionBlock(
-                                        factory.CodeTransition(),
-                                            factory.Code("DateTime.Now")
-                                                .AsImplicitExpression(CSharpCodeParser.DefaultKeywords)
-                                                .Accepts(AcceptedCharacters.NonWhiteSpace)));
-                var doWhileString = "@do { var foo = bar; <text>Foo</text> foo++; } while (foo<bar>);";
-                var doWhile = new MarkupBlock(
-                                new StatementBlock(
-                                    factory.CodeTransition(),
-                                    factory.Code("do { var foo = bar;").AsStatement(),
-                                    new MarkupBlock(
-                                        new MarkupTagBlock(
-                                            factory.MarkupTransition("<text>")),
-                                        factory.Markup("Foo").Accepts(AcceptedCharacters.None),
-                                        new MarkupTagBlock(
-                                            factory.MarkupTransition("</text>")),
-                                        factory.CodeMarkup(" ").With(new StatementChunkGenerator()).Accepts(AcceptedCharacters.None)),
-                                    factory.Code("foo++; } while (foo<bar>);").AsStatement().Accepts(AcceptedCharacters.None)));
-
-                var currentFormattedString = "<p class=\"{0}\" style='{0}'></p>";
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, dateTimeNowString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                        new List<KeyValuePair<string, SyntaxTreeNode>>
-                        {
-                            new KeyValuePair<string, SyntaxTreeNode>("class", new MarkupBlock(dateTimeNow)),
-                            new KeyValuePair<string, SyntaxTreeNode>("style", new MarkupBlock(dateTimeNow))
-                        }))
-                };
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, doWhileString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                        new List<KeyValuePair<string, SyntaxTreeNode>>
-                        {
-                            new KeyValuePair<string, SyntaxTreeNode>("class", new MarkupBlock(doWhile)),
-                            new KeyValuePair<string, SyntaxTreeNode>("style", new MarkupBlock(doWhile))
-                        }))
-                };
-
-                currentFormattedString = "<p class=\"{0}\" style='{0}'>Hello World</p>";
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, dateTimeNowString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", new MarkupBlock(dateTimeNow)),
-                                new KeyValuePair<string, SyntaxTreeNode>("style", new MarkupBlock(dateTimeNow))
-                            },
-                            factory.Markup("Hello World")))
-                };
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, doWhileString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", new MarkupBlock(doWhile)),
-                                new KeyValuePair<string, SyntaxTreeNode>("style", new MarkupBlock(doWhile))
-                            },
-                            factory.Markup("Hello World")))
-                };
-
-                currentFormattedString = "<p class=\"{0}\">Hello</p> <p style='{0}'>World</p>";
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, dateTimeNowString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", new MarkupBlock(dateTimeNow))
-                            },
-                            factory.Markup("Hello")),
-                        factory.Markup(" "),
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("style", new MarkupBlock(dateTimeNow))
-                            },
-                            factory.Markup("World")))
-                };
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, doWhileString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", new MarkupBlock(doWhile))
-                            },
-                            factory.Markup("Hello")),
-                        factory.Markup(" "),
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("style", new MarkupBlock(doWhile))
-                            },
-                            factory.Markup("World")))
-                };
-
-                currentFormattedString =
-                    "<p class=\"{0}\" style='{0}'>Hello World <strong class=\"{0}\">inside of strong tag</strong></p>";
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, dateTimeNowString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", new MarkupBlock(dateTimeNow)),
-                                new KeyValuePair<string, SyntaxTreeNode>("style", new MarkupBlock(dateTimeNow))
-                            },
-                            factory.Markup("Hello World "),
-                            new MarkupTagBlock(
-                                factory.Markup("<strong"),
-                                new MarkupBlock(new AttributeBlockChunkGenerator(name: "class",
-                                                                                prefix: new LocationTagged<string>(" class=\"", 66, 0, 66),
-                                                                                suffix: new LocationTagged<string>("\"", 87, 0, 87)),
-                                    factory.Markup(" class=\"").With(SpanChunkGenerator.Null),
-                                    new MarkupBlock(new DynamicAttributeBlockChunkGenerator(new LocationTagged<string>(string.Empty, 74, 0, 74), 74, 0, 74),
-                                        new ExpressionBlock(
-                                            factory.CodeTransition(),
-                                            factory.Code("DateTime.Now")
-                                                   .AsImplicitExpression(CSharpCodeParser.DefaultKeywords)
-                                                   .Accepts(AcceptedCharacters.NonWhiteSpace))),
-                                    factory.Markup("\"").With(SpanChunkGenerator.Null)),
-                                factory.Markup(">")),
-                            factory.Markup("inside of strong tag"),
-                            blockFactory.MarkupTagBlock("</strong>")))
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(ComplexAttributeTagHelperBlockData))]
-        public void TagHelperParseTreeRewriter_RewritesComplexAttributeTagHelperTagBlocks(
-            string documentContent,
-            MarkupBlock expectedOutput)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, "p");
-        }
-
-        public static IEnumerable<object[]> ComplexTagHelperBlockData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-                var blockFactory = new BlockFactory(factory);
-                var dateTimeNowString = "@DateTime.Now";
-                var dateTimeNow = new ExpressionBlock(
-                                        factory.CodeTransition(),
-                                            factory.Code("DateTime.Now")
-                                                .AsImplicitExpression(CSharpCodeParser.DefaultKeywords)
-                                                .Accepts(AcceptedCharacters.NonWhiteSpace));
-                var doWhileString = "@do { var foo = bar; <p>Foo</p> foo++; } while (foo<bar>);";
-                var doWhile = new StatementBlock(
-                               factory.CodeTransition(),
-                               factory.Code("do { var foo = bar;").AsStatement(),
-                               new MarkupBlock(
-                                    factory.Markup(" "),
-                                    new MarkupTagHelperBlock("p",
-                                        factory.Markup("Foo")),
-                                    factory.Markup(" ").Accepts(AcceptedCharacters.None)),
-                               factory.Code("foo++; } while (foo<bar>);").AsStatement().Accepts(AcceptedCharacters.None));
-
-                var currentFormattedString = "<p>{0}</p>";
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, dateTimeNowString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p", dateTimeNow))
-                };
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, doWhileString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p", doWhile))
-                };
-
-                currentFormattedString = "<p>Hello World {0}</p>";
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, dateTimeNowString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            factory.Markup("Hello World "),
-                            dateTimeNow))
-                };
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, doWhileString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            factory.Markup("Hello World "),
-                            doWhile))
-                };
-
-                currentFormattedString = "<p>{0}</p> <p>{0}</p>";
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, dateTimeNowString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p", dateTimeNow),
-                        factory.Markup(" "),
-                        new MarkupTagHelperBlock("p", dateTimeNow))
-                };
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, doWhileString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p", doWhile),
-                        factory.Markup(" "),
-                        new MarkupTagHelperBlock("p", doWhile))
-                };
-
-                currentFormattedString = "<p>Hello {0}<strong>inside of {0} strong tag</strong></p>";
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, dateTimeNowString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            factory.Markup("Hello "),
-                            dateTimeNow,
-                            blockFactory.MarkupTagBlock("<strong>"),
-                            factory.Markup("inside of "),
-                            dateTimeNow,
-                            factory.Markup(" strong tag"),
-                            blockFactory.MarkupTagBlock("</strong>")))
-                };
-                yield return new object[]
-                {
-                    string.Format(currentFormattedString, doWhileString),
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            factory.Markup("Hello "),
-                            doWhile,
-                            blockFactory.MarkupTagBlock("<strong>"),
-                            factory.Markup("inside of "),
-                            doWhile,
-                            factory.Markup(" strong tag"),
-                            blockFactory.MarkupTagBlock("</strong>")))
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(ComplexTagHelperBlockData))]
-        public void TagHelperParseTreeRewriter_RewritesComplexTagHelperTagBlocks(
-            string documentContent,
-            MarkupBlock expectedOutput)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, "p");
-        }
-
-        public static IEnumerable<object[]> ScriptBlockData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-
-                yield return new object[]
-                {
-                    "<script><script></foo></script>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("script",
-                            factory.Markup("<script></foo>")))
-                };
-                yield return new object[]
-                {
-                    "<script>Hello World <div></div></script>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("script",
-                            factory.Markup("Hello World <div></div>")))
-                };
-                yield return new object[]
-                {
-                    "<script>Hel<p>lo</p></script> <p><div>World</div></p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("script",
-                            factory.Markup("Hel<p>lo</p>")),
-                        factory.Markup(" "),
-                        new MarkupTagHelperBlock("p",
-                            new MarkupTagHelperBlock("div",
-                                factory.Markup("World"))))
-                };
-                yield return new object[]
-                {
-                    "<script>Hel<strong>lo</strong></script> <script><span>World</span></script>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("script",
-                            factory.Markup("Hel<strong>lo</strong>")),
-                        factory.Markup(" "),
-                        new MarkupTagHelperBlock("script",
-                            factory.Markup("<span>World</span>")))
-                };
-                yield return new object[]
-                {
-                    "<script class=\"foo\" style=\"color:red;\" />",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("script",
-                            selfClosing: true,
-                            attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                                new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                            }))
-                };
-                yield return new object[]
-                {
-                    "<p>Hello <script class=\"foo\" style=\"color:red;\"></script> World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            factory.Markup("Hello "),
-                            new MarkupTagHelperBlock("script",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                                    new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                                }),
-                            factory.Markup(" World")))
-                };
-                yield return new object[]
-                {
-                    "<p>Hello <script class=\"@@foo@bar.com\" style=\"color:red;\"></script> World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            factory.Markup("Hello "),
-                            new MarkupTagHelperBlock("script",
-                                new List<KeyValuePair<string, SyntaxTreeNode>>
-                                {
-                                    new KeyValuePair<string, SyntaxTreeNode>(
-                                        "class",
-                                        new MarkupBlock(
-                                            new MarkupBlock(
-                                                factory.Markup("@").Accepts(AcceptedCharacters.None),
-                                                factory.Markup("@").With(SpanChunkGenerator.Null).Accepts(AcceptedCharacters.None)),
-                                            factory.Markup("foo@bar.com"))),
-                                    new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                                }),
-                            factory.Markup(" World")))
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(ScriptBlockData))]
-        public void TagHelperParseTreeRewriter_RewritesScriptTagHelpers(
-            string documentContent,
-            MarkupBlock expectedOutput)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, "p", "div", "script");
-        }
-
-        public static IEnumerable<object[]> SelfClosingBlockData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-
-                yield return new object[]
-                {
-                    "<p class=\"foo\" style=\"color:red;\" />",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            selfClosing: true,
-                            attributes:  new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                                new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                            }))
-                };
-                yield return new object[]
-                {
-                    "<p>Hello <p class=\"foo\" style=\"color:red;\" /> World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock(
-                            "p",
-                            selfClosing: false,
-                            children: new SyntaxTreeNode[]
-                            {
-                                factory.Markup("Hello "),
-                                new MarkupTagHelperBlock(
-                                    "p",
-                                    selfClosing: true,
-                                    attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                                        {
-                                            new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                                            new KeyValuePair<string, SyntaxTreeNode>(
-                                                "style",
-                                                factory.Markup("color:red;"))
-                                        }),
-                                factory.Markup(" World")
-                            }))
-                };
-                yield return new object[]
-                {
-                    "Hello<p class=\"foo\" /> <p style=\"color:red;\" />World",
-                    new MarkupBlock(
-                        factory.Markup("Hello"),
-                        new MarkupTagHelperBlock("p",
-                            selfClosing: true,
-                            attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo"))
-                            }),
-                        factory.Markup(" "),
-                        new MarkupTagHelperBlock("p",
-                            selfClosing: true,
-                            attributes: new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                            }),
-                        factory.Markup("World"))
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(SelfClosingBlockData))]
-        public void TagHelperParseTreeRewriter_RewritesSelfClosingTagHelpers(
-            string documentContent,
-            MarkupBlock expectedOutput)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, "p");
-        }
-
-        public static IEnumerable<object[]> QuotelessAttributeBlockData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-                var blockFactory = new BlockFactory(factory);
-                var dateTimeNow = new MarkupBlock(
-                                    new ExpressionBlock(
-                                        factory.CodeTransition(),
-                                            factory.Code("DateTime.Now")
-                                                .AsImplicitExpression(CSharpCodeParser.DefaultKeywords)
-                                                .Accepts(AcceptedCharacters.NonWhiteSpace)));
-
-                yield return new object[]
-                {
-                    "<p class=foo dynamic=@DateTime.Now style=color:red;></p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                        new List<KeyValuePair<string, SyntaxTreeNode>>
-                        {
-                            new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                            new KeyValuePair<string, SyntaxTreeNode>("dynamic", new MarkupBlock(dateTimeNow)),
-                            new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                        }))
-                };
-                yield return new object[]
-                {
-                    "<p class=foo dynamic=@DateTime.Now style=color:red;>Hello World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                                new KeyValuePair<string, SyntaxTreeNode>("dynamic", new MarkupBlock(dateTimeNow)),
-                                new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                            },
-                            factory.Markup("Hello World")))
-                };
-                yield return new object[]
-                {
-                    "<p class=foo dynamic=@DateTime.Now style=color@@:red;>Hello World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                                new KeyValuePair<string, SyntaxTreeNode>("dynamic", new MarkupBlock(dateTimeNow)),
-                                new KeyValuePair<string, SyntaxTreeNode>(
-                                    "style",
-                                    new MarkupBlock(
-                                        factory.Markup("color"),
-                                         new MarkupBlock(
-                                            factory.Markup("@").Accepts(AcceptedCharacters.None),
-                                            factory.Markup("@").With(SpanChunkGenerator.Null).Accepts(AcceptedCharacters.None)),
-                                        factory.Markup(":red;")))
-                            },
-                            factory.Markup("Hello World")))
-                };
-                yield return new object[]
-                {
-                    "<p class=foo dynamic=@DateTime.Now>Hello</p> <p style=color:red; dynamic=@DateTime.Now>World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                                new KeyValuePair<string, SyntaxTreeNode>("dynamic", new MarkupBlock(dateTimeNow))
-                            },
-                            factory.Markup("Hello")),
-                        factory.Markup(" "),
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;")),
-                                new KeyValuePair<string, SyntaxTreeNode>("dynamic", new MarkupBlock(dateTimeNow))
-                            },
-                            factory.Markup("World")))
-                };
-                yield return new object[]
-                {
-                    "<p class=foo dynamic=@DateTime.Now style=color:red;>Hello World <strong class=\"foo\">inside of strong tag</strong></p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                                new KeyValuePair<string, SyntaxTreeNode>("dynamic", new MarkupBlock(dateTimeNow)),
-                                new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                            },
-                            factory.Markup("Hello World "),
-                            new MarkupTagBlock(
-                                factory.Markup("<strong"),
-                                new MarkupBlock(new AttributeBlockChunkGenerator(name: "class",
-                                                                                prefix: new LocationTagged<string>(" class=\"", 71, 0, 71),
-                                                                                suffix: new LocationTagged<string>("\"", 82, 0, 82)),
-                                    factory.Markup(" class=\"").With(SpanChunkGenerator.Null),
-                                    factory.Markup("foo").With(new LiteralAttributeChunkGenerator(prefix: new LocationTagged<string>(string.Empty, 79, 0, 79),
-                                                                                                 value: new LocationTagged<string>("foo", 79, 0, 79))),
-                                    factory.Markup("\"").With(SpanChunkGenerator.Null)),
-                                factory.Markup(">")),
-                            factory.Markup("inside of strong tag"),
-                            blockFactory.MarkupTagBlock("</strong>")))
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(QuotelessAttributeBlockData))]
-        public void TagHelperParseTreeRewriter_RewritesTagHelpersWithQuotelessAttributes(
-            string documentContent,
-            MarkupBlock expectedOutput)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, "p");
-        }
-
-        public static IEnumerable<object[]> PlainAttributeBlockData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-                var blockFactory = new BlockFactory(factory);
-
-                yield return new object[]
-                {
-                    "<p class=\"foo\" style=\"color:red;\"></p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                        new List<KeyValuePair<string, SyntaxTreeNode>>
-                        {
-                            new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                            new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                        }))
-                };
-                yield return new object[]
-                {
-                    "<p class=\"foo\" style=\"color:red;\">Hello World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                                new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                            },
-                            factory.Markup("Hello World")))
-                };
-                yield return new object[]
-                {
-                    "<p class=\"foo\">Hello</p> <p style=\"color:red;\">World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo"))
-                            },
-                            factory.Markup("Hello")),
-                        factory.Markup(" "),
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                            },
-                            factory.Markup("World")))
-                };
-                yield return new object[]
-                {
-                    "<p class=\"foo\" style=\"color:red;\">Hello World <strong class=\"foo\">inside of strong tag</strong></p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            new List<KeyValuePair<string, SyntaxTreeNode>>
-                            {
-                                new KeyValuePair<string, SyntaxTreeNode>("class", factory.Markup("foo")),
-                                new KeyValuePair<string, SyntaxTreeNode>("style", factory.Markup("color:red;"))
-                            },
-                            factory.Markup("Hello World "),
-                            new MarkupTagBlock(
-                                factory.Markup("<strong"),
-                                new MarkupBlock(new AttributeBlockChunkGenerator(name: "class",
-                                                                                prefix: new LocationTagged<string>(" class=\"", 53, 0, 53),
-                                                                                suffix: new LocationTagged<string>("\"", 64, 0, 64)),
-                                    factory.Markup(" class=\"").With(SpanChunkGenerator.Null),
-                                    factory.Markup("foo").With(new LiteralAttributeChunkGenerator(prefix: new LocationTagged<string>(string.Empty, 61, 0, 61),
-                                                                                                 value: new LocationTagged<string>("foo", 61, 0, 61))),
-                                    factory.Markup("\"").With(SpanChunkGenerator.Null)),
-                                factory.Markup(">")),
-                            factory.Markup("inside of strong tag"),
-                            blockFactory.MarkupTagBlock("</strong>")))
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(PlainAttributeBlockData))]
-        public void TagHelperParseTreeRewriter_RewritesTagHelpersWithPlainAttributes(
-            string documentContent,
-            MarkupBlock expectedOutput)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, "p");
-        }
-
-        public static IEnumerable<object[]> PlainBlockData
-        {
-            get
-            {
-                var factory = CreateDefaultSpanFactory();
-                var blockFactory = new BlockFactory(factory);
-
-                yield return new object[]
-                {
-                    "<p></p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p"))
-                };
-                yield return new object[]
-                {
-                    "<p>Hello World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            factory.Markup("Hello World")))
-                };
-                yield return new object[]
-                {
-                    "<p>Hello</p> <p>World</p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            factory.Markup("Hello")),
-                        factory.Markup(" "),
-                        new MarkupTagHelperBlock("p",
-                            factory.Markup("World")))
-                };
-                yield return new object[]
-                {
-                    "<p>Hello World <strong>inside of strong tag</strong></p>",
-                    new MarkupBlock(
-                        new MarkupTagHelperBlock("p",
-                            factory.Markup("Hello World "),
-                            blockFactory.MarkupTagBlock("<strong>"),
-                            factory.Markup("inside of strong tag"),
-                            blockFactory.MarkupTagBlock("</strong>")))
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(PlainBlockData))]
-        public void TagHelperParseTreeRewriter_RewritesPlainTagHelperTagBlocks(
-            string documentContent,
-            MarkupBlock expectedOutput)
-        {
-            RunParseTreeRewriterTest(documentContent, expectedOutput, "p");
         }
 
         public static IEnumerable<object[]> NestedBlockData
